+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   declaration.
 * In interactive mode, make spacebar also force refresh
 * Update punctuation cleanup to work around more script types (like Chinese)
+* Make interactive view support unicode
 
 ### 0.1.8
 * Support set forks with `{$id: (foo), (bar)}` syntax. Silent set