+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * In interactive mode, make spacebar also force refresh
 * Update punctuation cleanup to work around more script types (like Chinese)
 * Make interactive view support unicode
+* Fix interactive view confusingly suppressing error messages, for
+  example when trying to read a non-existent file.
 
 ### 0.1.8
 * Support set forks with `{$id: (foo), (bar)}` syntax. Silent set